Refactor maxDose calculation to simplify conversion logic and ensure proper rounding in smartRound function

+++ lib/utils/smart_rounder.dart
@@ -14,7 +14,7 @@ num smartRound(double value) {
 
   if (digits >= 3) {
     double factor = pow(10, digits - 3).toDouble();
-    return ( (value / factor).round() * factor ).toInt();
+    return ( (value / factor).round() * factor ).toInt().toDouble();
   } else {
     int decimalPlaces = 3 - digits;
     double roundedValue = (value * pow(10, decimalPlaces)).round() / pow(10, decimalPlaces);
